% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{base_uris}
\alias{base_uris}
\title{base_uris}
\format{
A data frame:
\describe{
\item{retailer_name}{Name of the energy retailer}
\item{retailer_base_uri}{URI for API}}
\item{cdr_brand}{URI name for API}}
}
\usage{
base_uris
}
\description{
CDR Base URI data from https://www.aer.gov.au/energy-product-reference-data
}
\keyword{datasets}
