% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_api.R
\name{call_eprd_api}
\alias{call_eprd_api}
\title{Download, extract, and tidy AER Energy plan metadata}
\usage{
call_eprd_api(
  api_uri,
  offer_type = "all",
  fuel_type = "all",
  effective = "all"
)
}
\arguments{
\item{api_uri}{character. Supply retailer uniform resource identifier.}

\item{offer_type}{character; default is "all". Other values are "standing",
"market" and "regulated"}

\item{fuel_type}{character; default is "all". Other values are "gas",
"electricity" and "dual"}

\item{effective}{character; default is "all". Other values are "current" and
"future"}
}
\description{
Download, extract, and tidy AER Energy plan metadata
}
