% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_api_plandetail.R
\name{call_eprd_api_plandetail}
\alias{call_eprd_api_plandetail}
\title{Call EPRD API plan detail}
\usage{
call_eprd_api_plandetail(base_uri, planid)
}
\arguments{
\item{base_uri}{character. A retailer cdr brand name. See \code{readeprd::base_uris} for valid names.}

\item{planid}{character; Supply an EME or VEFS offer identifier}
}
\description{
Call EPRD API plan detail
}
