% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eprd.R
\name{read_eprd}
\alias{read_eprd}
\title{Download, extract, and tidy energy plan data
#'}
\usage{
read_eprd(retailer = NULL, planid = "all", fuel_type = "all", source = "vec")
}
\arguments{
\item{retailer}{character. A retailer cdr brand name. See \code{readeprd::base_uris} for valid names.}

\item{planid}{(optional) character; default is "all". Supply an EME or VEFS offer
identifier (such as \code{TAN617921MR@VEC}) to get only that plan.}

\item{fuel_type}{character; default is "all". Other values are "electricity" and "gas".}

\item{source}{character; default is "vec". Choose plans from Victorian Energy Compare or Energy Made Easy.
Other values are "eme" and "all".}
}
\value{
data
}
\description{
Download, extract, and tidy energy plan data
#'
}
