% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eprd_metadata.R
\name{read_eprd_metadata}
\alias{read_eprd_metadata}
\title{Reads retailer plan metadata in a single data frame}
\usage{
read_eprd_metadata(retailer = "all", fuel_type = "all")
}
\value{
data
}
\description{
#' @param retailer character; default is "all". A retailer cdr brand name.
See \code{readeprd::base_uris} for valid names.
#' @param fuel_type character; default is "all". Other values are "electricity"
and "gas".
}
