% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plandetail.R
\name{read_eprd_plan}
\alias{read_eprd_plan}
\title{nest contract details into the plan id}
\usage{
read_eprd_plan(base_uri, planid)
}
\arguments{
\item{base_uri}{character. A retailer cdr brand name. See \code{readeprd::base_uris} for valid names.}

\item{planid}{character; Supply an EME or VEFS offer identifier}
}
\description{
nest contract details into the plan id
}
