% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readeprd-package.R
\docType{package}
\name{readeprd-package}
\alias{readeprd}
\alias{readeprd-package}
\title{readeprd: Read Energy Product Reference Data from Consumer Data Rights API}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Miguel Ayala \email{miguel.ayala@esc.vic.gov.au}

}
\keyword{internal}
